requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
playwright>=1.45,<2.0
pdfplumber>=0.11,<1.0
pdfminer.six>=20231228
